requirements.txt
asgiref==3.6.0
Django==4.1.6
django-environ==0.9.0
django-filter==22.1
django-pytest==0.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
